actions: fix how fault response tables get attached to plan
So far, #attach_to was called by the fault response table's initialize
method. However, we create a new table object each time we instanciate
an action, as e.g. when some fault response is triggered. This
was causing the table to be attached multiple times to the plan
(and, eventually, hell).

Fix this by removing the call to #attach_to from the table's initialize,
and making Plan#use_fault_response_table call it. The way interface-level
fault tables get attached is also reworked (done by Application instead
of Interface#initialize)