coord: fix handling of instruction whose #execute method returns false
The general assumption is that {#execute} is called only once and that
the rest is done through event handlers. However, if for whatever
reason #step gets called with a pending instruction, no validation
was done to make sure that the current instruction could be skipped.

Force "long-running" instructions to call #cancel before they call
{#step}