coord: FaultHandler.replace_by should remove the dependency to the faulty task before terminating
ReplaceBy keeps the dependency relations of the response location
for (1) keeping track of what is happending and (2) ensuring that
the fault will still be there if the fault handler fails.

However, once the fault handler is successful, we really have to
make sure that this dependency relation is removed (otherwise,
the original fault is raised again).