test: fix some execution engine test that were locking
The issue was in a change in Test#process_events, namely that we
now wait one cycle when the engine is running. The offending tests
were making the engine *believe* that it was running (while it was
not), which was making #process_events block.

The calls to Test#process_events have been replaced with a direct
call to engine.process_events in these tests, with appropriate
explanations in the comments