coord: implement a way to get access to the action state and action script elements by name
Action state machines and scripts get named subtasks based on the
assignation of the task objects to local variables. Moreover, we
automatically assign the name of the action when a script uses the
blabla! syntax.

This now allows to resolve particular states / script children by
name through Coordination::Models::Base.