app: implement an auto-load-all vs. auto-load-local logic
This is an idea that originally comes from syskit. Currently, the
model-loading logic was loading everything that it could find
in #search_path, which is counter-productive if we want to
have "model repositories" apps, i.e. apps where only parts of
the models are actually applicable to the real application.

This introduces the auto_load_all? flag to control whether we
should only autoload from the current app or from search_path.
The flag is false by default.