test: convert to minitest and cleanup common test setup
Minitest is the de-facto testing library that we are targetting
(because of its inclusion as the ruby testing library in ruby).
However, the minitest gem is sometimes incompatible with Ruby's
test/unit shim. This leads to being unable to run the test suite.
Convert to the minitest API completely to avoid these problems.

Moreover, this cleans up the common test setup and the test files
to remove unnecessary include and require statements.