core: create a fastpath for plan task/event insertion
Plan#add is currently very costly, as it tries to discover
the whole subplan of the task/events that are getting added.
While we can't do away with it right now, this commit adds
simpler methods that don't do the discovery, but instead only
add the task to the plan.

At this stage, this is for the benefit of some tooling (e.g.
Syskit) which don't need the costly discovery step.