droby: refactor link-reestablishment logic
One front-facing change is the name, changed from 'reconnect' to
'reestablish_link'. The former was confusing as peer-to-peer
"connections" are meant as a state of cooperation, not a state of
communication.

The rest of the code was modified to use as little blocking calls
as possible, to avoid deadlocks, and do some cleanup of the reconnection
logic as well