test: modify minitest's exception handling to account for Roby peculiarities
Namely, we now remove the default exception display notifications during
an assert_raise (as the assert_raise will display the exception properly),
use #pretty_print to display exceptions (since it is what Roby expects),
and make sure that we unwrap non-Roby exceptions if the assert_raise points
to a non-Roby exception. Meaning that

   assert_raise(RuntimeError) { ... }

will work fine even if the exception got wrapped in e.g. a CodeError