test: disable the scheduler by default in process_events_synchronous
This method is really used only for tests, where one wants to emit
or call a specific event directly. The scheduler interfers with the
nature of the test by adding other events to be emitted at the same
time (and, also, probably slows down the tests altogether).

This makes the scheduling optional, and disables it by default