log: tremendously speed up initialization time when connecting to a log server
Log::Client#read_and_process_pending limits the amount of data it
processes in order to avoid blocking too long to one chunk. However,
the callers were ignoring that (i.e. calling it once per Qt timer),
which led to really wasting a lot of time during initialization as
well as sometime when something complex was happening.

This keeps the do-not-block-too-long policy, but this time by specifying
an acceptable amount of time the method can block and trying to block
at most that long.