core: one more attempt at rationalizing exception reporting
This is one more attempt at making the reporting of exceptions more
rational. The main reason is the ability to properly report Roby's
very common "composite exceptions" in a GUI (as opposed to as plain
text in the console).

With this change:
 - all Roby exceptions may define a #original_exceptions accessor
   that enumerates the exceptions that caused the receiver to be
   raised itself.
 - pretty_print does NOT display any of the original exceptions
   anymore, only toplevel display methods do so
 - pretty_print does NOT display any backtrace anymore, only toplevel
   display methods do so

The change required us to create a Roby-specific subclass of
MetaRuby::GUI::ExceptionView, that should be used in GUIs that want
to properly display roby exceptions