log: inform clients on the initialization progress
During initialization, we basically process a lot of log data as fast
as possible. To speedup the process, one should not update any display
during that phase, and only start displaying afterwards.

This commit makes it possible by informing the clients about the init
progress, both about its progress itself as well as when it ended.