interface: remove the RECAPTURED state and some other corner-case fixes
RECAPTURED is now handled as a MONITORED notification. Moreover, REPLACED
if followed by a MONITORED/<state notification> pair.

The constraint that add_mission / add_permanent trigger with the
mission/permanent flag has been relaxed. This makes sure we have a
single code path between the to-be-retired #add_mission and
#add_permanent and the new #add_mission_task and #add_permanent_task.
That constraint does not make a lot of sense in the end as adding
a task as mission can also be separated between a call to #add
and a call to #add_mission, so making the two code paths equivalent
ensures consistency (same for #add_permanent)