interface: stop processing in #poll as soon as we get a cycle_end message
This ensures that one can do some synchronous per-cycle processing if
one wishes to.

The biggest change is in the way we wait for data. Since we might have some data
already read that contains full websocket frames, we cannot simply IO.select
anymore but must actively try to parse the already-read data. It actually
simplifies the logic in Client and in DRobyChannel.