core: do parallel assignment of arguments in #initialize and #freeze_delayed_arguments
It is common for subclasses to define a high-level argument that is,
in the end, propagated to lower-level arguments (see test for an
example). This commit fixes a corner case of using this scheme, where
one gives both the high level and low level argument values at the same
time. This happens e.g. during log replay, and currently fails because
arguments are single-assignation.