core: refactor the relation graph architecture
This commit refactors the relation system to allow for each plan
having its own set of graph objects (instead of sharing the same).

It overhauls completely the test suite for the relation code, but
leaves everything else broken as-is. Necessary changes in the rest
of the Roby code will follow