core: refactor plan and graph handling to have one set of graphs per plan
This commit modifies the graph handling away from having one global graph for
all plans, instead having one set of graph objects per plan.

The biggest change is that the graph API is now not available anymore on the
global objects (e.g. TaskStructure::Dependency). In the few places where such
access was required, the graph must now be resolved explicitely. Moreover
having `object.plan == nil` now does not indicate that the task is "new", but
that it has been removed.

In Plan, the direct benefit is that we don't have to "discover" a new task's
neighborhood anymore (its plan is holding it). All plan objects are now created
with a Roby::TemplatePlan which holds their internal structure. When tasks that
have not been explicitely added to a non-template plan are joined, their plans
get merged. Finally, when Plan#add is called the tasks template plan is merged
into the "work plan"

In Transaction, during commit, the transaction gets modified to keep only the
relations between new tasks, which is then merge!'d into the underlying plan.
The rest of the relations are then added to the underlying plan as usual.

To make this happen, the application of relation modifications
involving proxies has been moved out of the plan object's
commit_transaction. #commit_transaction should from now on only
deal with modifications internal to the plan object. Inter-object
modifications are handled by the relation handling and Transaction
itself