core: replace our BGL wrapper by a RGL-compatible graph class
The wrapper is actually a performance hog. Not BGL itself, but the
back-and-forth between BGL data structures and Ruby, and the fact
that we can't optimize some things without either implementing them
in C++ or changing the BGL structure tremendously.

Finally, the usage of a C++ wrappers makes the whole relation graphs
and task objects "immune" to generational GC.

Algorithmically speaking, RGL offers all we need. Just transition to it.