benchmark: add benchmark of passing an enumeration block with yield() or as a block object
This is meant to figure out whether it is faster to delegate enumeration
by having an explicit
   each { |v| yield(v) }

or by using the &block syntax

   def delegating_each(&block)
      object.each(&block)
   end

TL;DR: the call to delegating_each is slower in the second form,
but the enumeration is faster. When doing both, it seems that the
second form wins in most cases where actually does enumerate
something.