core: fundamentally modify the relation hook handling
Hooks are now only emitted on an ExecutablePlan on template plans and
transactions, it is assumed that validation will be done globally "when
it is best" (which, obviously, in the case of the transactions, is
at least at commit time).

The hooks are a horrible performance hog, and in practice really provides
very little - the only problematic place so far is the handling
of AndGenerator and OrGenerator, which could be easily fixed in
merged_plan.