log: rationalize logging and hook handling
This commit primary objective is to make sure that logging works with the new
relation structure and operations. However, in addition, it cleans up the list
of log messages as well as how these messages are handled.

Most hook methods are removed. The 'super if defined? super' idion is removed
since Ruby now has Module#prepend. Logging messages are emitted right at the
relevant point in the core class.

Finally, we now track only the main executable plan.  Tracking transactions is
way too costly, and in addition would track TemplatePlan (which we most probably
don't want). The support to track transactions will probably be re-added in the
future, though.