core: don't forcefully wrap task event generators when a task is wrapped
This wrapping is usually useless - there's very little modifications to
the event net - and very costly since we get to wrap and then compare
the "standard" task event nets.

The biggest change / problem this induced is in subplan replacement.
Since we want to avoid loading the whole network in the transaction,
we must do the replacement comparisons on the merged relations.