droby: rewrite the whole droby protocol and marshalling/demarshalling
This work stems from the need to dump plan objects, required to get
the Roby 3 data structures properly logged. This cleanup was
required for a long LONG time so ...

This rips out the distributed namespace and tests, which are broken
for years, and won't be worked on in the near future. If we end up rewriting it,
then we'll do some git archeology.

The namespace survived so far because it was holding the marshalling/
demarshalling code, needed for logging or shell. This code is now moved
into its own namespace and significantly cleaned up. Moreover, the
marshalling structures are now versioned (bumping the current version
to V5) to guarantee that we will be able to upgrade log files in the
future.