droby: make the reporting of scheduler state incremental
So far, the scheduler state was sent whole just before the scheduler
was called. Which makes no sense because (1) the cycle N does not
contain all the information related to itself and (2) the tasks are
not marshalled at the point where the message is generated (this is
how I noticed the problem, the new log code cannot marshal
already-finalized tasks).

Change to report each scheduler report as it arrives, aggregating them
on the plan rebuilder side.