core: replace the old #waiting_threads functionality by something compatible with Concurrent::Obligation
Any object that respond to #fail and #complete? can now be registered
in a #waiting thread-safe array. #fail will be called on engine quit,
and #complete? is used to garbage-collect the array.

It replaces the old waiting_threads mechanism (which was using the
Thread#raise method).

EE#once has gained a 'sync' argument to register an ivar used to
synchronize the calling thread with the once block. It's set by
EE#execute and #wait_until