core: refactor ExecutionEngine#garbage_collect
This is a piece of the Roby infrastructure that has not been touched
for years, and is in dear need of a face lift. The two main goals of
this refactoring (apart from having a properly tested GC implementation)
are:

1. stop tasks that are only needed by finished tasks

This makes sure that keeping a subplan alive with a relation does NOT
keep the tasks alive forever. The current remove_when_done flag for
depends_on was originally implemented because of this limitation.

2. allow to specify that a finished task should either keep its
subplan alive, or should be garbage collected (finished & its subplan
removed), but not finalized. This is to give more control when using
Roby's plan for state tracking. For instance, it would allow Syskit
to remove the workarounds when having to disconnect tasks that are
not present in the Syskit plan anymore.