app: allocate ports outside the ephemeral port ranges of the major OSes
This is a bug that I noticed for a long time, but could not pinpoint.
We would end up with the Roby shell server port being taken by what
looks like a weird self-connected socket (i.e. localhost:4242 to
localhost:4242). This happens to be valid TCP, and happens when the
random ephemeral port allocated for the client socket matches the
server's, when the client is created before the server is.

It ends up creating a socket that is connected to itself. Apart
from the port being taken (if a Roby app gets created in the meantime),
another symptom was the wrong Marshal format - because the client
sends something and gets to read it back, which it did not expect to do.