core: fix incremental updates of SchedulingConstraints#task_graph on edge removal
With the change in using a glob pattern in the Rakefile to run
the tests, I discovered that test/relations/test_event_relation_graph.rb
was not run (or maybe never run), and was obviously broken after
the code got moved to the SchedulingConstraint relation.

This commit moves the tests where it belongs, and finally fixes the
task_graph maintenance code w.r.t. incremental updates on removal.
An issue that I have the feeling I already "discovered" 10 times.