core: give control over finalization with PlanObject#can_finalize?
It does not give control over garbage-collection (which will still
happen), but forbid the engine to remove the task. It is useful when
some asynchronous post-stop cleanup or pre-start configuration has to
be done (as is the case in Syskit with resp. connection removal or
task configuration). In these cases, it is useful to keep the task
in the plan to allow to keep track of these synchronous actions more
easily.