core: refactor promise support to avoid context switches
Roby::Promise is not designed to distribute work, but to avoid blocking
the main thread when external calls block (e.g. remote calls, I/O).

Concurrent::Promise is a very competent promise implementation, but it
requeues each promise in the thread pool. In Roby's case it is very
wasteful as apart from sleeping, the promises are expected to do very
little.

This commit refactors Promise to handle the processing pipeline itself,
to avoid switching threads and to make sure that a sequence of in-engine
handlers are called in one go. It reduces the waiting time a lot, with
very little effect on the main promise role (avoid blocking).