test: update #process_events and add #process_events_until
ExecutionEngine#process_events now returns the set of errors that
happened during this cycle. Use this return value instead of capturing
the errors through the #on_exception hook.

The method also got changed so that we process all waiting work
iteratively.

Moreover, add #process_events_until that ensures that tests are immune
to the number of calls to process_events