coordination: extract CoordinationAction from Actions::Models::MethodAction
Because of laziness, MethodAction was handling both "normal" method
actions, and actions backed by a coordination model. With the complexity
and difficult codepath that ensued.

This commit splits the two, making the presence/absence of a coordination
model explicit, and moving the instanciation of a coordination model
into the action object's #instanciate method instead of as a method
defined on the action interface. The method still exists of course for
the benefit of method-actions.

This required to change how the Action descriptions are created. They
now accept another Action as argument, which they use for initialization.
This allows to use a plain Action in #describe, which then gets promoted
to a more specific action model when we get to know what model it is
(e.g. MethodAction or CoordinationAction)

Because of this change, I discovered that Coordination::Models::Base
had very weird semantics regarding #find_child (it was returning
states, which are not children as they are not always active in
coordination models). I couldn't find a good reason for it in the
commit history, so it got removed.