actions: do not override an existing action in #use_library
This is in line with Syskit's policy with profiles (and use_profile),
and is so far the one with the least surprising (what one has to do
is explicitely load the basic actions/profiles first, which is really
what one does naturally).

The better fix would be to avoid importing actions/Syskit definitions
recursively (i.e. have use_library only import the actions that have
been defined in that particular library)