relations: replace the internal graph structure by hashes that have compare_by_identity set
Graph access and manipulation is currently the single most expensive
low-level post in the graph manipulation, and hash access is responsible
for that in a major way.

We store only objects that can be compared by identity in the BD graph.
However, the current structure was using array pairs as keys, which we
cannot do anymore. The result is actually cleaner.

This yields a 30% improvement in the modify-plan-with-transactions
benchmark.