coordination: fix a few nasty bugs in action state machine transition
The test titles speak for themselves, so let's describe the one
problem that is not explicitely part of the tests - but is triggered
by both of them.

When a state reuses an existing task, in some cases it is possible that
the state's success event is already emitted (as is the 'intermediate'
event in one of the two tests in this commit). In that case, the
dependency would simply be ignored (and with remove_when_done, removed).

This commit therefore introduces a #from_now predicate on single event
predicates that ignore any event emitted before that time. It ensures
that the dependency follows the "natural" temporal order.