core: move the check for errors out of #pending? into the schedulers
This check was meant to avoid scheduling tasks that are pending and
schedulable, but for which there is already an error (i.e. Roby is
trying to kill them). However, this has a bad interaction with the
garbage collector if the task is actually really pending.

Move the check to the scheduler where it really belongs