core: change how errors caught during error propagation are re-propagated themselves
These are now propagated iteratively within the event loop, instead
of having a custom additional error gathering in the error propagation
itself.

MissionFailedError and PermanentTaskFailed are now generated globally
instead of through handlers. This is necessary in order to avoid
generating more errors than we should. For instance, in a network of:

  RootA -> Child
  RootB -> Child

where 'Child' is a mission, we would get four MissionFailedError:
 - one created when the global handler hits the original error in RootA
 - one created when the global handler hits the original error in RootB

and then x2 because they are propagated through the network.

With this commit, only one MissionFailedError is generated, and then
x2 because of the propagation.