actions: fix actions from actions interface being registered on use'd libraries
With
   Child < Parent
and
   Child.use_library Library

then doing Child.an_action_from_parent was causing 'Library', and
Roby::Actions::Library itself, to get a 'an_action_from_parent'.
This led to basically importing the action on all the action
interfaces in the system.

This was caused by the presence of a promotion method on the module
common to Interface and Library. This module splits the definitions
so that only the interface gets the promotion method.