core: refactor exception propagation
These commits refactor the exception handling code. The seminal change was to
replace ExecutionEngine#trace - which is currently an array - by a full graph,
and reaping the benefits of it.

The main client-visible changes are:
 - the plan-level handlers are given a single exception object, regardless of
   the number of roots involved in the propagation. This is made possible thanks
   to the ability to merge all unhandled exceptions
 - handled exceptions are now part of the phase result object
 - removed the 'propagated' flag. One can instead call add_error and set the
   propagate-through parents to an empty array to achieve the same result
   without a specific flag

The code now exfiltrates handled and inhibited error information, mostly for the
benefit of the testing environment, but one never knows.