core: remove the 'do not schedule if there is a queued exception' clause
This was a problem with (much) older version of the event propagation
loop, where the last step in the loop was to propagate the exceptions,
but the scheduler was running before Roby had a change to act on it.

This is not the case anymore, so the GC will remove any pending toplevel
task *before* the scheduler has a chance to see them.