coordination: implement a way to capture an event context into a state machine variable
It regularly happens that values that are measured/determined/...
by a state need to be propagated (usually as an argument) to another
state. This is currently done "behind the scenes", usually through
the State global object.

A workaround was to pass the information through the parent task, and
then use from(:parent_task).<accessorname>. However, this is fragile
as it assumes that there is only one parent task.

This commit implements a way to store the information from an event into
a state-machine-local variable, which can then be used as argument into
other states.