droby: cleanup the handling of registration of peer IDs in log files
The current way registration was handled was hacked together "to
make it work". Which of course means that it does not.

It now follows the overall intended design. The mapping from local
ID to "remote" (that is, log file) ID is done explicitly by the
logger. On the other side, the PlanRebuilder is modified to also
explicitely handle registration from the log-side IDs ("remote")
to the rebuilder-local IDs ("local")