core: fix schedulers considering failed-to-start tasks in a loop
emit_failed is called by the exception handling, which itself is
called after the event processing, a.k.a. after scheduling gets
nothing to do.

So, we basically get
  scheduler calls Task.start
  Task.start raises, pending is reset and error is queue
  scheduler calls Task.start again

This ensures that failed_to_start! is set directly from the
emit_failed handler, and adds a test for the infinite loop
directly.