core: refactor model-level argument definition to add documentation strings
The current argument implementation was relying on two different
inherited hashes, one for default values and one for names. This
consolidates it into a single attribute that holds an Argument
struct.

It also cleans the default argument handling to use a null object for
"no default", instead of relying on complex logic