all the test_depend are actually depend_optional
The GUI and development tools (syskit ide, tests, ...) depends on
all the test dependencies to run. Moreover, stackprof and rbtrace
are actually needed for the --debug shell interface.

Make them optional, as they are not strictly needed, but do install
them by default as a lot of expected functionality depend on them