droby: unconditional unmarshal properties that may hold droby-identifiable objects
This is the sibling of cee3ccdb0d478d6780cf4d7a3e10f52dc6409425.

When sending droby-identifiable objects the first time, the sending
side assumes that said object will be identifiable on the remote side
fron now on.

However, whenever we proxy an object by mapping it (by e.g. name) to
an existing object on the remote side, we often were not doing anything
to the properties held by it. For instance, a MethodAction return type
would not be registered on the remote side if the MethodAction could
already be resolved by name.

This led to the infamous "cannot find object for DRobyID ..." error.

This was triggered by a bug report, but I did a review of all the
proxy() implementations to make sure that there is no other obvious
cases.