tasks: refactor ExternalProcess to enable the :pipe and :close redirection modes
This refactors ExternalProcess to allow redirecting the process' output
to a method on the task (resp. stdout_received and stderr_received).
This mode is enabled by passing :pipe to #redirect_output, e.g.

    task.redirect_output(stdout: :pipe)

Moreover, it adds the :close mode, which closes said output, that is

   task.redirect_output(stdout: :close)

Finally, it provides the ability to redirect to existing files (as e.g.
/dev/null), which was impossible before

   task.redirect_output('/dev/null')