cli: reset the INT trap handler to DEFAULT on `run`
The generally expected behavior of `roby run` depends on INT
being passed to the Ruby runtime. Without this, the display server
and Syskit's process server won't quit as expected

It so happens that an IGNORED signal is not reset on exec, so
if syskit's parent process ignores INT, this breaks Syskit. This
actually happens in VSCode.