test: fix huge performance hit due to lstat calls during teardown
The app teardown rm_rf's the log_dir, because the logs are private.
However, *because* the private flag was set after the log dir creation,
the log dir contains the 'current' symlink.

It seems to cost quite a bit given the number of times we do it.
Fixing this removes ~35% of the run time of the whole test suite,
from 237 to 154s on my machine.