interface: rework how errors are handled server-side
Connections are now always closed on protocol error. This includes
when a reply fails to be marshalled (i.e. we can't notify the remote
side that a reply succeeded). A new message, :protocol_error, is
also used to notify the remote side of the reason for the close.

The optional behavior change controlled by abort_on_exception? in
Server has been removed, as this class is too deep to actually be
able to abort anything. The implementation for this is instead moved
to TCPServer, which has knowledge about the application and the engine.