app: fix corner-cases in the app teardown path under error
ExecutionEngine#run fails the pending waiting work when it quits.
Because it's actually a race, this can fail (i.e. we .fail(), but the work
actually finished itself in-between) and concurrent-ruby raises then.

When this happened, Application#join would basically misbehave
and go into an infinite loop.

This reviews and unit tests the overall codepath. Moreoover, it
adds a warning about waiting work being terminated, as it should
not happen (we should have proper synchronization that avoids
this)