interface: make sure UI event notifications are MT-safe
UI events are meant to be sent from various places in the system,
and are really meant to be forwarded to the outside interfaces.
It's not reasonable to require that they are all generated from
the event loop.

This fixes an actual bug where file watching in Syskit would send
an UI event (conf file change) which would then end up writing packets
in the wrong thread.