interface: fix race between exception reporting and finalization
The exception queue and the job queue are separate. This led to having
job monitors being removed *before* we had the chance to process the
exceptions.

Make sure we finalize after the exceptions have been processed. Note
that we still must process the job state notifications first, as
this is where we actually create the monitors in the first place.