interface: ensure async gets notifications only after it got the job set
The current async connection would connect (which calls #handshake),
and then call .jobs on the connected client object. From a timeline
point of view, it means that the server would start sending
notifications for existing jobs before the client resolves the jobs
call. If jobs die during this execution cycle, this leads to having
notifications on jobs that were not part of the received jobs set.

The code currently deals with this gracefully, but there's no reason
to keep this weird timeline. Make the call to 'jobs' part of the
handshake, and ensure that no notifications are sent until the
handshake has been processed.

This also removes the synchronization when establishing callbacks
using execution_engine.execute. This was not needed anymore - for
quite a while - since all calls are processed within the event
loop on the server side.