interface: always send JOB_FINALIZED to the remote end
Remote job tracking keeps track of a job even in the DROPPED state,
in order to be able to recover it if needed. It therefore does require
getting JOB_FINALIZED

This is triggered when a job is dropped, but is finalized in a following
cycle because either its own stop command is asynchronous, or because
it is kept alive for a while.