droby: fix plan rebuild for garbaged but not finalized tasks
The finalization of garbage tasks can be delayed if their
can_finalize? handler returns true. Something that e.g. Syskit
uses to asynchronously clean up before it loses the task object.

The plan rebuilder would still clean those in #clear_integrated,
which would lead the task to be removed twice (once in the first
garbage call, the other in the garbage event that actually
finalizes the task)