app: fix deleting created log base directories on 2.5
When a tool like the IDE quits, it attempts to delete all
directories that it had to create for its own log dir. However,
if in the meantime something like syskit run was executed, the
log directories end up not being empty (and we definitely don't
want to delete them).

On Ruby 2.3, FileUtils.rmdir would simply ignore such directories.
2.5 raises Errno::ENOTEMPTY. Since Errno::ENOTEMPTY already exists
in 2.3 as well, let's just rescue it.