cli: whitelist methods that should be called in the main thread
The current method is not good enough since some (fairly critical)
calls do access the remote channel (e.g. #process).

Whitelist the methods that need to be called in the main thread,
which are few, and add a few more that are obviously safe purely
for performance reasons. This should be safer long-term.