shell: add a timeout on calls
This ensures that client such as the shell or the IDE don't freeze
forever if Syskit has become unresponsive

I was originally going for a timeout: parameter, but it is tricky to
make the difference between a `timeout` argument to a remote call and a
timeout argument to the Client method. It would have needed a different
design from the get-go. So, go for a class-wide timeout instead.