test: provide Test::InstanceExtension to allow filtering on robot/mode/... even on non-Roby contexts
The roby_should_run mechanism, which is the underlying implementation
for e.g. run_interactive was currently available only on Roby/Syskit-
related tests, as e.g. composition tests or task context tests.

This makes it available also for "plain" tests. The only requirement
is that the test needs to extend Roby::Test::DSL.