test: fix test error that would happen only sporadically
Ruby does not seem to like having a "hidden" IO duplicate of
a given file descriptor the way the --fd test does. This was
causing a EBADF in some test orderings.

Explicitly close the underlying I/O object, and ignore the
error that is caused by the fact CLI::Display already closed
it.