queries: generalize plan-iteration capabilities to all matchers
So far, the Query class was the only one that could be used to
find tasks matching a TaskMatcher inside a plan. With the
recent addition of all types of matchers (events, task events,
and/or/not) and the need to add a port matcher in Syskit, this
commit refactors the code to allow any type of matcher to iterate
over a plan.

The main change in the commit is to replace the `query_set_result`/
`query_each` delegation from Query to plans by specific classes
with a defined interface (#include? / #each_in_plan). The old
method was expecting the result of query_set_result to be an
opaque type that could only be iterated upon by calling #query_each
on the plan that generated it.

This allowed me to relatively easily rewrite the #roots operator on
transactions to deal with any depth of stacked transactions (still very
expensive operation, use with care)