interface: interpret EADDRNOTAVAIL as a ConnectionError
Ruby's internal implementation will try other hosts if it gets
an ECONNREFUSED. For instance, if one sets 127.0.0.1 **and** ::1
as 'localhost', it will try to contact `::1` after 127.0.0.1 returned
ECONNREFUSED.

Unfortunately, in one very important environment for us (CI),
Kubernetes forcefully sets localhost to both 127.0.0.1 and ::1 but
does not allow for IPv6 resolution inside the container. This
leads the ::1 attempt to return EADDRNOTAVAIL and **that** is the
error reported by Ruby.

For the time being, assume EADDRNOTAVAIL is a ConnectionError.
The worst case would be to take longer (timeout) to actually fail.