dump timepoints only for cycles that are longer than the expected cycle length
Timepoint logging is actually a huge overhead in size - it is by far
the majority of the size of a roby event log for long-running Roby
apps.

In the end, it is essentially needed only to debug longer-than-expected
cycles. Dump them only for those cycles, and get rid of them for
the rest.