core: remove duplicate tags when merging fullfilled models
I hit a bug where two Syskit would get merged in a way that
the same fullfille model, which had tags, was merged over and
over again. Exponentials ... Who likes those ?

Anyways, the tag list is supposed to be small, so I preferred
keeping the array and calling uniq! afterwards on it