cli: explicitly require 'readline' in shell
Shell-related test started failing in CI because Readline was
not defined. My guess is that we depended on a library that
was requiring 'readline' and that library stopped doing so.

In any case, this pointed out that we're using inconsistent
readline implementations. IRB uses Readline while we are using
RbReadline. Probably not a good idea, but since I have little time
and since it works so far, I'm leaving it as-is for now.